% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_totals.R
\name{drop_totals}
\alias{drop_totals}
\title{drop_totals}
\usage{
drop_totals(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Drops rows containing totals
}
\details{
Takes a data.frame and filters it to return rows that do not contain the
word "total" without paying attention to case.
}
\author{
Alyssa Savo
}
