% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clarity.R
\name{read_clarity}
\alias{read_clarity}
\title{read_clarity}
\usage{
read_clarity(url, destfile, Web01 = FALSE, report = NULL)
}
\arguments{
\item{destfile}{A path and filename to save the .zip file to.}

\item{Web01}{Clarity Elections sites that use Web01 formatting, which include
Web01 in the URL, may not include direct links to .zip reports. In this case,
set `Web01` to TRUE and the function will attempt to construct a link to the report.}

\item{report}{This should be NULL unless Web01 is TRUE. Choose "csv" for a summary
report or "xls", "xlm", or "txt" for detailed reports in those formats.}

\item{file}{The URL to a zipped report on a Clarity Elections website, or the
homepage of a Web01 Clarity site.}
}
\description{
Downloads election results using the Clarity Elections platform.
}
\details{
Automatically downloads and unzips election reports from elections websites that
use Scytl's Clarity Elections platform. Users can either provide a link directly
to the .zip file or, for Web01 Clarity sites (noted in the website URL), link
to the home page and specify the desired report.
}
\author{
Alyssa Savo
}
