% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results}
\alias{read_results}
\title{read_results}
\usage{
read_results(file)
}
\arguments{
\item{file}{A URL or path to a .pdf file.}
}
\description{
Imports and tidies election results from simply-formatted .pdf files.
}
\details{
This function should help read and tidy election results from simple and well-
formatted .pdf files, where localities (county, city, voting district, etc.)
are listed in the first column and candidate vote totals are listed in all
subsequent columns.
}
\author{
Alyssa Savo
}
