% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_totals.R
\name{get_totals}
\alias{get_totals}
\title{get_totals}
\usage{
get_totals(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Retrieves rows containing totals
}
\details{
Takes a data.frame and filters it to return only the rows that contain the
word "total" without paying attention to case.
}
\examples{
essex_gov_17 <- read.csv("data/essex_gov_17.csv")
drop_totals(essex_gov_17)

}
\author{
Alyssa Savo
}
