% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_twocol_summary.R
\name{read_twocol_summary}
\alias{read_twocol_summary}
\title{read_twocol_summary}
\usage{
read_twocol_summary(file)
}
\arguments{
\item{file}{A file or URL linking containing a .pdf file containing election results
in a standard two-column summary format.}
}
\description{
Imports and tidies election results data from two-column summary reports.
}
\details{
Downloads and tidies .pdf files containing election results in a standardized
two-column summary format.
}
\author{
Alyssa Savo
}
