% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vertical_results.R
\name{read_vertical_results}
\alias{read_vertical_results}
\title{read_vertical_results}
\usage{
read_vertical_results(file, range, colnames)
}
\arguments{
\item{file}{A URL or path to a .pdf file.}

\item{range}{An integer vector specifying the pages in the PDF for the race of interest. There currently is not a way to read multiple races at once.}

\item{colnames}{A character vector containing the names for each column.}
}
\description{
Reads election results from .pdf files where column titles are formatted vertically.
}
\details{
Vertically formatted text is often formatted incorrectly when converting
.pdf files to tabular data formats such as data.frames and tibbles. This
function includes additional parameters to help generate well-formatted
data.frames for a single race from .pdfs containing election results with
vertical column headers.
}
\examples{
file <- "data/dem_primary_essex_17.pdf"
read_vertical_results(file, range = c(1:11), colnames = c("Municipality","Registration","Ballots Cast","Turnout (\%)","Philip MURPHY","William BRENNAN","John S. WISNIEWSKI","Jim Johnson","Mark ZINNA","Raymond J. LESNIAK","Write-In"))

}
\author{
Alyssa Savo
}
